supress regexp/use-ignore-case errors
This rule recommends to simplify the following usage

/^[dDsSwW]/ to /^[dsw]/i

This regex is for matching the CharacterClassEscape

// CharacterClassEscape :: one of d D s S w W

If we simplify to the /i flag and suppose CharacterClassEscape now allows a new production `x` but not `X`, it is prone to errors that we accidentally allows `X` because of the `i` flag. For this reason I suggest to disable this rule.